refactor: centralize useAuth hook and roles export
