+++ src/features/auth/components/LoginForm.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useAuth } from '../../../contexts/AuthContext';
+import useAuth from '../../../hooks/useAuth';
 import styles from '../auth.module.css';
 import { useNavigate, useLocation } from 'react-router-dom';
