+++ src/features/auth/components/SignupForm.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useAuth } from '../../../contexts/AuthContext';
+import useAuth from '../../../hooks/useAuth';
 import styles from '../auth.module.css';
 import { useNavigate } from 'react-router-dom';
