+++ src/features/dashboard/AdminPage.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { useAuth } from '../../contexts/AuthContext';
+import useAuth from '../../hooks/useAuth';
 // Note: usersApi import is commented out until backend implementation is ready
 // import { usersApi } from '../../lib/apiClient';
