+++ src/features/home/Home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useAuth } from '../../contexts/AuthContext';
+import useAuth from '../../hooks/useAuth';
 import WaveBackground from './components/WaveBackground';
 import KpiCard from './components/KpiCard';
 import styles from '../../styles/home.module.css';
