+++ src/features/plantas/PlantasPage.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { useAuth } from '../../contexts/AuthContext';
+import useAuth from '../../hooks/useAuth';
 import { plantasApi } from '../../lib/apiClient';
 
 export default function PlantasPage() {
