+++ src/router/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
+import useAuth from "../hooks/useAuth";
 
 export default function ProtectedRoute({ roles, children }) {
   const { isAuthenticated, user } = useAuth();
